move forcing tests to separate file
